Fixes #103 Refactor map_loader.rs into map_loader/ submodule
